Refactor ArgumentParser, use toml instead of config, and move to cli
