+++ pyproject.toml
@@ -56,7 +56,7 @@ packages = ["isimip_utils"]
 version = { attr = "isimip_utils.__version__" }
 
 [tool.ruff]
-target-version = "py39"
+target-version = "py311"
 line-length = 120
 select = [
     "B",    # flake8-bugbear
